Renaming .env in e2e tests to .env-e2e
