+++ Meadowlark-js/tests/e2e/setup/Teardown.ts
@@ -3,7 +3,8 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-const environment = require('./EnvironmentConfig');
+// @ts-ignore
+const environment = require('./TestEnvironmentSetup');
 
 async function endServer() {
   try {
